fix: create common module for fix endpoints
Create common module on the fix API module to allow
other endpoints and CLI to reuse functions
